% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bcca.R
\name{bcca}
\alias{bcca}
\alias{dbcca}
\alias{pbcca}
\alias{qbcca}
\alias{rbcca}
\title{The Box-Cox Cauchy Distribution}
\usage{
dbcca(x, mu, sigma, lambda, log = FALSE, ...)

pbcca(q, mu, sigma, lambda, lower.tail = TRUE, ...)

qbcca(p, mu, sigma, lambda, lower.tail = TRUE, ...)

rbcca(n, mu, sigma, lambda)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{vector of strictly positive scale parameters.}

\item{sigma}{vector of strictly positive relative dispersion parameters.}

\item{lambda}{vector of real-valued skewness parameters. If \code{lambda = 0}, the Box-Cox
Cauchy distribution reduces to the log-Cauchy distribution with parameters
\code{mu} and \code{sigma} (see \code{\link{lca}}).}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{...}{further arguments.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dbcca} returns the density, \code{pbcca} gives the distribution function,
    \code{qbcca} gives the quantile function, and \code{rbcca} generates random observations.

    Invalid arguments will result in return value \code{NaN}, with a warning.

    The length of the result is determined by \code{n} for \code{rbcca}, and is the
    maximum of the lengths of the numerical arguments for the other functions.
}
\description{
Density, distribution function, quantile function, and random generation for the Box-Cox Cauchy
    distribution with parameters \code{mu}, \code{sigma}, and \code{lambda}.
}
\examples{
mu <- 10
sigma <- 0.5
lambda <- 4

# Sample generation
x <- rbcca(10000, mu, sigma, lambda)

# Density
hist(x, prob = TRUE, main = "The Box-Cox Cauchy Distribution", col = "white")
curve(dbcca(x, mu, sigma, lambda), add = TRUE, col = 2, lwd = 2)
legend("topright", "Probability density function", col = 2, lwd = 2, lty = 1)

# Distribution function
plot(ecdf(x), main = "The Box-Cox Cauchy Distribution", ylab = "Distribution function")
curve(pbcca(x, mu, sigma, lambda), add = TRUE, col = 2, lwd = 2)
legend("bottomright", c("Emp. distribution function", "Theo. distribution function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)

# Quantile function
plot(seq(0.01, 0.99, 0.001), quantile(x, seq(0.01, 0.99, 0.001)),
  type = "l",
  xlab = "p", ylab = "Quantile function", main = "The Box-Cox Cauchy Distribution"
)
curve(qbcca(x, mu, sigma, lambda), add = TRUE, col = 2, lwd = 2, from = 0, to = 1)
legend("topleft", c("Emp. quantile function", "Theo. quantile function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
    applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.

Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical
    properties and parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
