% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bchp.R
\name{bchp}
\alias{bchp}
\alias{dbchp}
\alias{pbchp}
\alias{qbchp}
\alias{rbchp}
\title{The Box-Cox Hyperbolic Distribution}
\usage{
dbchp(x, mu, sigma, lambda, nu, log = FALSE)

pbchp(q, mu, sigma, lambda, nu, lower.tail = TRUE)

qbchp(p, mu, sigma, lambda, nu, lower.tail = TRUE)

rbchp(n, mu, sigma, lambda, nu)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{vector of strictly positive scale parameters.}

\item{sigma}{vector of strictly positive relative dispersion parameters.}

\item{lambda}{vector of real-valued skewness parameters. If \code{lambda = 0}, the Box-Cox
hyperbolic distribution reduces to the log-hyperbolic distribution with parameters
\code{mu}, \code{sigma}, and \code{nu} (see \code{\link{lhp}}.}

\item{nu}{strictly positive heavy-tailness parameter.}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dbchp} returns the density, \code{pbchp} gives the distribution function,
    \code{qbchp} gives the quantile function, and \code{rbchp} generates random observations.

    Invalid arguments will result in return value \code{NaN}, with an warning.

    The length of the result is determined by \code{n} for \code{rbchp}, and is the
    maximum of the lengths of the numerical arguments for the other functions.
}
\description{
Density, distribution function, quantile function, and random
    generation for the Box-Cox hyperbolic distribution with parameters \code{mu},
    \code{sigma}, \code{lambda}, and \code{nu}.
}
\examples{
mu <- 8
sigma <- 1
lambda <- 2
nu <- 4

# Sample generation
x <- rbchp(10000, mu, sigma, lambda, nu)

# Density
hist(x, prob = TRUE, main = "The Box-Cox Hyperbolic Distribution", col = "white")
curve(dbchp(x, mu, sigma, lambda, nu), add = TRUE, col = 2, lwd = 2)
legend("topleft", "Probability density function", col = 2, lwd = 2, lty = 1)

# Distribution function
plot(ecdf(x), main = "The Box-Cox Hyperbolic Distribution", ylab = "Distribution function")
curve(pbchp(x, mu, sigma, lambda, nu), add = TRUE, col = 2, lwd = 2)
legend("bottomright", c("Emp. distribution function", "Theo. distribution function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)

# Quantile function
plot(seq(0.01, 0.99, 0.001), quantile(x, seq(0.01, 0.99, 0.001)),
  type = "l",
  xlab = "p", ylab = "Quantile function", main = "The Box-Cox Hyperbolic Distribution"
)
curve(qbchp(x, mu, sigma, lambda, nu), add = TRUE, col = 2, lwd = 2, from = 0, to = 1)
legend("topleft", c("Emp. quantile function", "Theo. quantile function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)
}
\references{
Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical
    properties and parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.

Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
    applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
