% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bcloi.R
\name{bcloi}
\alias{bcloi}
\alias{dbcloi}
\alias{pbcloi}
\alias{qbcloi}
\alias{rbcloi}
\title{The Box-Cox Type I Logistic Distribution}
\usage{
dbcloi(x, mu, sigma, lambda, log = FALSE, ...)

pbcloi(q, mu, sigma, lambda, lower.tail = TRUE, ...)

qbcloi(p, mu, sigma, lambda, lower.tail = TRUE, ...)

rbcloi(n, mu, sigma, lambda)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{vector of strictly positive scale parameters.}

\item{sigma}{vector of strictly positive relative dispersion parameters.}

\item{lambda}{vector of real-valued skewness parameters. If \code{lambda = 0}, the Box-Cox
type I logistic distribution reduces to the log-type I logistic distribution with parameters
\code{mu} and \code{sigma} (see \code{\link{lloi}}).}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{...}{further arguments.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dbcloi} returns the density, \code{pbcloi} gives the distribution function,
    \code{qbcloi} gives the quantile function, and \code{rbcloi} generates random observations.

    Invalid arguments will result in return value \code{NaN}, with an warning.

    The length of the result is determined by \code{n} for \code{rbcloi}, and is the
    maximum of the lengths of the numerical arguments for the other functions.
}
\description{
Density, distribution function, quantile function, and random
    generation for the Box-Cox type I logistic distribution with parameters \code{mu},
    \code{sigma}, and \code{lambda}.
}
\examples{
mu <- 8
sigma <- 1
lambda <- 2

# Sample generation
x <- rbcloi(10000, mu, sigma, lambda)

# Density
hist(x, prob = TRUE, main = "The Box-Cox Type I Logistic Distribution", col = "white")
curve(dbcloi(x, mu, sigma, lambda), add = TRUE, col = 2, lwd = 2)
legend("topleft", "Prob. density function", col = 2, lwd = 2, lty = 1)

# Distribution function
plot(ecdf(x), main = "The Box-Cox Type I Logistic Distribution", ylab = "Distribution function")
curve(pbcloi(x, mu, sigma, lambda), add = TRUE, col = 2, lwd = 2)
legend("bottomright", c("Emp. distribution function", "Theo. distribution function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)

# Quantile function
plot(seq(0.01, 0.99, 0.001), quantile(x, seq(0.01, 0.99, 0.001)),
  type = "l",
  xlab = "p", ylab = "Quantile function", main = "The Box-Cox Type I Logistic Distribution"
)
curve(qbcloi(x, mu, sigma, lambda), add = TRUE, col = 2, lwd = 2, from = 0, to = 1)
legend("topleft", c("Emp. quantile function", "Theo. quantile function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
    applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.

Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical
    properties and parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
