% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_choose_copula.R
\name{choose_copula-methods}
\alias{choose_copula-methods}
\alias{print.choose_copula}
\alias{plot.choose_copula}
\title{Methods for 'choose_copula' objects}
\usage{
\method{print}{choose_copula}(x, ...)

\method{plot}{choose_copula}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"choose_copula"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Methods for 'choose_copula' objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
