% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_choose_copula.R
\name{choose_copula}
\alias{choose_copula}
\title{Choose the Extra Parameter of the BCS-NSM Marginal Regression}
\usage{
choose_copula(
  object,
  grid = seq(0.5, 0.98, 0.04),
  copula,
  trace = TRUE,
  plot = TRUE,
  control = control_fit(...),
  ...
)
}
\arguments{
\item{object}{a \code{"marginreg"} object.}

\item{grid}{grid of values that will be used to evaluate the profiled log-likelihood function.
If the tweaks are not computationally intensive, we suggest \code{grid = 1:15}.}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the NSM copula. Currently,
the copulas available are: Gaussian (\code{"gaussian"}),
Student's t (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{trace}{logical; if \code{TRUE}, a summary with the profiled log-likelihood value, the AIC,
the BIC, and the run-time of the fit is displayed.}

\item{plot}{logical; if \code{TRUE}, a graph of the profiled log-likelihood evaluated in the
considered grid of values is shown.}

\item{control}{a list of control arguments specified via \code{\link{control_fit}}.}

\item{...}{further arguments passed to \code{\link{control_fit}}.}
}
\value{
An object of class \code{"choose_copula"}. More specifically, it returns a list in which
    each element consists of the fit of a BCS-NSM marginal regression with each value of the extra
    parameter specified in \code{grid}.
}
\description{
Estimation of the extra parameter of the NSM copula in the fit of the BCS-NSM marginal regression
via profile log-likelihood.
}
\references{
Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical properties and
    parameter estimation. *Brazilian Journal of Probability and Statistics*, 30, 196-220.

Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and applications to
    nutritional data. *AStA Advances in Statistical Analysis*, 101, 321-344.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
