% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_control.R
\name{control_fit}
\alias{control_fit}
\title{Optimization Control Parameters Passed to Optim}
\usage{
control_fit(method = "BFGS", maxit = 2000, hessian = TRUE, inits = NULL, ...)
}
\arguments{
\item{method}{the method to be used. See "Details" in \code{\link[stats]{optim}}. The default
method (\code{"BFGS"}) is a quasi-Newton method (also known as a variable metric algorithm),
specifically that published simultaneously in 1970 by Broyden, Fletcher, Goldfarb and Shanno.}

\item{maxit}{the maximum number of iterations of the algorithm. Defaults to \code{2000}.}

\item{hessian}{logical. Should a numerically differentiated Hessian matrix be returned?}

\item{inits}{an optional vector with starting values for all parameters for fitting a BCNSM
distribution. It must be passed in the order: \code{(mu, sigma, lambda, nu, gamma)}, where
\code{mu}, \code{sigma}, \code{lambda}, and \code{nu} are parameters associated with marginal
distributions and \code{gamma} are parameters related to the association matrix.}

\item{...}{further arguments to be passed to \code{\link[stats]{optim}}.}
}
\value{
A list with the arguments specified.
}
\description{
Optimization parameters passed to \code{\link[stats]{optim}} for the fit of a BCS-NSM marginal
    regression model via \code{\link{marginreg}}. This function acts
    in the same spirit as \code{\link[betareg]{betareg.control}} from the \code{betareg} package.
    Its primary purpose is to gather all the optimization control arguments in a single function.
}
\references{
Cribari-Neto, F., and Zeileis, A. (2010). Beta regression in R.
    \emph{Journal of statistical software}, 34, 1-24.

Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical properties and
    parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.

Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and applications to
    nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
