% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bct.R
\name{lt}
\alias{lt}
\alias{dlt}
\alias{plt}
\alias{qlt}
\alias{rlt}
\title{The Log-t Distribution}
\usage{
dlt(x, mu, sigma, nu, log = FALSE, ...)

plt(q, mu, sigma, nu, lower.tail = TRUE, ...)

qlt(p, mu, sigma, nu, lower.tail = TRUE, ...)

rlt(n, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{vector of strictly positive scale parameters.}

\item{sigma}{vector of strictly positive relative dispersion parameters.}

\item{nu}{strictly positive heavy-tailedness parameter.}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{...}{further arguments.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dlt} returns the density, \code{plt} gives the distribution
    function, \code{qlt} gives the quantile function, and \code{rlt}
    generates random observations.

    Invalid arguments will result in return value \code{NaN}.

    The length of the result is determined by \code{n} for \code{rlt}, and is the
    maximum of the lengths of the numerical arguments for the other functions.
}
\description{
Density, distribution function, quantile function, and random
    generation for the log-t distribution with parameters \code{mu},
    \code{sigma}, and \code{nu}.
}
\details{
A random variable X has a log-t distribution with parameter \code{mu} and
    \code{sigma} if log(X) follows a Student's t distribution with location parameter \code{log(mu)}
    and dispersion parameter \code{sigma}. It can be showed that \code{mu} is the median of X.
}
\examples{
mu <- 8
sigma <- 1
nu <- 4

# Sample generation
x <- rlt(10000, mu, sigma, nu)

# Density
hist(x, prob = TRUE, main = "The Log-t Distribution", col = "white")
curve(dlt(x, mu, sigma, nu), add = TRUE, col = 2, lwd = 2)
legend("topright", "Probability density function", col = 2, lwd = 2, lty = 1)

# Distribution function
plot(ecdf(x), main = "The Log-t Distribution", ylab = "Distribution function")
curve(plt(x, mu, sigma, nu), add = TRUE, col = 2, lwd = 2)
legend("bottomright", c("Emp. distribution function", "Theo. distribution function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)

# Quantile function
plot(seq(0.01, 0.99, 0.001), quantile(x, seq(0.01, 0.99, 0.001)),
  type = "l",
  xlab = "p", ylab = "Quantile function", main = "The Log-t Distribution"
)
curve(qlt(x, mu, sigma, nu), add = TRUE, col = 2, lwd = 2, from = 0, to = 1)
legend("topleft", c("Emp. quantile function", "Theo. quantile function"),
  col = c(1, 2), lwd = 2, lty = c(1, 1)
)
}
\references{
Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical
    properties and parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
