% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fit_function.R
\name{marginreg}
\alias{marginreg}
\title{BCS-NSM Marginal Regression Models for Positive Continuous Data}
\usage{
marginreg(
  formula,
  data,
  subset,
  na.action,
  margin = "bcno",
  mu.link = "log",
  sigma.link = "log",
  association = nonassociative(),
  copula = "gaussian",
  eta = NULL,
  control = control_fit(...),
  ...
)
}
\arguments{
\item{formula}{a simbolic description of the model, of type
\code{y ~ x} for covariates in the mean submodel only or \code{y ~ x | z}
to enter covariates in the variance submodel. See details below.}

\item{data}{an optional data frame containing the variables in the formula. By default the
variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting
process. (See additional details about how this argument interacts with data-dependent bases
in the ‘Details’ section of the \code{\link[stats]{model.frame}} documentation.)}

\item{na.action}{a function which indicates what should happen when the data contain \code{NAs}.
The default is set by the \code{na.action} setting of \link[base]{options}, and is
\code{\link[stats]{na.fail}} if that is unset. The ‘factory-fresh’ default is
\code{\link[stats]{na.omit}}. Another possible value is \code{NULL}, no action.
Value \code{\link[stats]{na.exclude}} can be useful.}

\item{margin}{a character which specifies the marginal Box-Cox symmetric (BCS) distribution.
A table with the current available BCS distributions can be seen in \code{\link{bcs}}.}

\item{mu.link, sigma.link}{character; specifies the link function for the \code{mu} and
\code{sigma} submodels, respectively. The links \code{"log"} (default) and
\code{"identity"} are currently available.}

\item{association}{a function which returns an object of class \code{"\link{association}"},
that describes the dependence structure of the data. See \code{\link{association}}.}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the copula. Currently,
the copulas available are: Gaussian (\code{"gaussian"}),
Student's t (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{eta}{possible extra parameter induced by the copula, parameterized to take values at
\code{(0, 1)}. For instance, under the \emph{t} copula, \code{eta} specifies
\code{eta} / (1 - \code{eta}) degrees of freedom. To estimate the value of \code{eta}
via profile log-likelihood, use the function \code{\link{choose_copula}}.}

\item{control}{a list of optimization control arguments specified via \code{\link{control_fit}}.}

\item{...}{further optimization control arguments passed to \code{\link{control_fit}}.}
}
\value{
The \code{marginreg} function returns an object of class \code{"marginreg"},
 which consists of a list with the following components:
\describe{
  \item{coefficients}{a list containing the elements \code{"beta"} and \code{"kappa"}.
      which consists of the estimates of the regression coefficients associated with the
      \code{mu} and the \code{sigma} submodels, respectively.}
  \item{fitted.values}{a list with the fitted values for \code{mu}, \code{sigma},
      \code{lambda}, and \code{nu}.}
  \item{margin}{a character with the marginal BCS distribution of the fit.}
  \item{links}{a list with elements \code{"mu.link"} and \code{"sigma.link"} with the specified link
      functions for the \code{mu} and \code{sigma} submodels, respectively.}
  \item{copula, eta}{\code{"copula"} is a character which informs which normal scale mixture distribution
      was used to generate the NSM copula and \code{"eta"} is the possible extra parameter associated with
      the copula, parameterized to take values at (0, 1).}
  \item{gamma}{the estimated parameters of the association matrix, if any.}
  \item{association}{structure of the association matrix. It can be one of \code{"non-associative"},
      \code{"unstructured"}, or \code{"uniform"}.}
  \item{logLik}{log-likelihood of the fitted model.}
  \item{vcov}{asymptotic covariance matrix of the maximum likelihood estimator of the model parameters vector.
      Specifically, the inverse of the observed information matrix, obtained via numeric Hessian matrix.}
  \item{y}{the response vector.}
  \item{X, Z}{the model matrices from the \code{mu} and \code{sigma} submodels, respectively.}
  \item{optim_params}{control optimization parameters used by \code{\link{control_fit}}.}
  \item{nobs,d}{the number of observations in the sample and the dimension of the response variable, respectively.}
  \item{call}{ the function call.}
  \item{formula}{the formula used to specify the model in \code{marginreg}.}
  \item{terms}{the terms for the "\code{mu}" submodel.}
 }
}
\description{
Fit a BCS-NSM marginal regression model to a possible correlated positive data with a Box-Cox
    symmetric distribution, where the dependence structure is described by a normal scale
    mixture copula.
}
\details{
The \code{marginreg} function implements a broad class of marginal regression models for
    analyzing correlated positive data with Box-Cox symmetric marginal distributions.
    The dependence is described by a normal scale mixture copula, offering a fully-parametric
    alternative to the classical generalized estimating equation models. The proposed class
    provides a flexible modeling framework for positive data with different levels of skewness
    and tail-heaviness, allowing for different dependence structures, including independence,
    time series, longitudinal, clustered, or spatially correlated data through the use of
    the copula.

The normal scale mixture copula entirely describes the dependence structure of the
    models. Because copulas are invariant under strictly increasing transformations,
    copula-based measures of association of the scale mixtures of normal distributions are the
    same for the BCS-NSM distributions. This statement is valid for measures of concordance, such
    as Kendall's tau or the tail dependence coefficients.

    Because copulas are invariant under strictly increasing transformations,
    copula-based measures of association of the scale mixtures of normal distributions are the
    same for the BCS-NSM distributions. This statement is valid for measures of concordance, such
    as Kendall's tau or the tail dependence coefficients. Let \eqn{\gamma_{ij}} be the element of
    the \emph{i}th row and \emph{j}th column of the association matrix \eqn{\bold{\Gamma}}, where
    \eqn{\gamma_{ii} = 1}, for \emph{i, j = 1, ..., d}. Thus, Kendall's tau of \eqn{Y_i} and 
    \eqn{Y_j} is given by

    \deqn{\tau(Y_i, Y_j) = \dfrac{2 \; \textrm{asin}(\gamma_{ij})}{\pi},}

    and it is invariant under the copula. Since Kendall’s tau is a
    rank correlation coefficient, it measures any monotonous association between two random variables.
    Thus, we have an increasing monotonous association between \eqn{Y_i} and \eqn{Y_j} when \eqn{\gamma_{ij}}
    is close to \eqn{1}, and a decreasing monotonous association when \eqn{\gamma_{ij}} is close to \eqn{-1}.
    Furthermore, \eqn{\gamma_{ij} = 0} indicates no monotonous association. These facts show that
    the association matrix \eqn{\bold{\Gamma}} plays an essential role in specifying the dependence
    structure of \eqn{\bold{Y}}. It can be parameterized in terms of a real-valued vector \eqn{\gamma}
    in an interpretative way depending on the type of association that is of interest to describe.
    
    Currently, the \code{bcsnsm.mr} package provides the non-associative, ARMA(p, q), cluster, and
    matern association structures, see \code{\link{association}}.
    
    The \code{formula} argument provides a symbolic description of the regression structure of
    the model following the implementation of beta regression in the \code{betareg} package
    (Cribari-Neto and Zeileis, 2010). For instance, the basic formula is of type \code{y ~ x1 + x2},
    which specifies the regression setting for the \code{mu} submodel. Additionally, a regression
    structure for the \code{sigma} submodel, say expressed in terms of \code{z1} and \code{z2}, 
    is specified as \code{y ~ x1 + x2 | z1 + z2} using functionalities from the \code{Formula}
    package (Zeileis and Croissant, 2010).
}
\references{
Cribari-Neto F, Zeileis A (2010). Beta regression in R. \emph{Journal
 of Statistical Software}, \bold{34}, 1–24

 Ferrari, S. L., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.
}
